Use type int when initializing SharedMemory size

+++ streaming/base/shared/array.py
@@ -24,7 +24,7 @@ class SharedArray:
         self.shape = np.empty(shape).shape
         self.dtype = dtype
         self.name = name
-        size = np.prod(shape) * dtype(0).nbytes
+        size = int(np.prod(shape) * dtype(0).nbytes)
         self.shm = SharedMemory(name=name, size=size)
 
     def numpy(self) -> NDArray:
